+++ src/views/products/list.vue
@@ -21,7 +21,7 @@
     <el-table :data="productsList" v-loading="listLoading" element-loading-text="加载中..." border fit highlight-current-row>
       <el-table-column width="65" align="center" label="序号" >
         <template slot-scope="scope">
-          <span>{{scope.$index+1}}</span>
+          <span>{{scope.row.id}}</span>
         </template>
       </el-table-column>
       <el-table-column width="200" align="center" label="产品名称" >
@@ -132,15 +132,26 @@ export default {
       this.$router.push({ name: 'editProduct', params: {id: id}})
     },
     handleDelete(id) {
-      deleteProduct(id).then(res => {
-        this.$notify({
-          title: '成功',
-          message: '删除成功',
-          type: 'success',
-          duration: 2000
+      this.$confirm('是否确定删除?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        deleteProduct(id).then(res => {
+          this.$notify({
+            title: '成功',
+            message: '删除成功',
+            type: 'success',
+            duration: 2000
+          })
+          this.getProductsList()
         })
-        this.getProductsList()
-      })
+      }).catch(() => {
+        this.$message({
+          type: 'info',
+          message: '已取消删除'
+        });
+      });
     }
   },
   mounted() {
